Linux variant of reserve_memory_special_concurrent to GTest
